+++ ruby_event_store/spec/actions/delete_stream_events_spec.rb
@@ -2,24 +2,19 @@ require_relative '../spec_helper'
 
 module RubyEventStore
   describe Facade do
-
-    let(:repository)  { InMemoryRepository.new }
-    let(:facade)      { RubyEventStore::Facade.new(repository) }
     let(:stream_name) { 'stream_name' }
     let(:page_size)   { 100 }
 
-    before(:each) do
-      repository.reset!
-    end
-
     specify 'raise exception if stream name is incorrect' do
+      facade = RubyEventStore::Facade.new(InMemoryRepository.new)
       expect { facade.delete_stream(nil) }.to raise_error(IncorrectStreamData)
       expect { facade.delete_stream('') }.to raise_error(IncorrectStreamData)
     end
 
     specify 'successfully delete streams of events' do
-      prepare_events_in_store('test_1')
-      prepare_events_in_store('test_2')
+      facade = RubyEventStore::Facade.new(InMemoryRepository.new)
+      prepare_events_in_store(facade, 'test_1')
+      prepare_events_in_store(facade, 'test_2')
       all_events = facade.read_all_streams_forward(:head, page_size)
       expect(all_events.length).to eq 8
       facade.delete_stream('test_2')
@@ -30,15 +25,10 @@ module RubyEventStore
 
     private
 
-    def prepare_events_in_store(stream_name)
+    def prepare_events_in_store(facade, stream_name)
       4.times do |index|
-        event = OrderCreated.new(event_id: index)
-        create_event(event, stream_name)
+        facade.publish_event(OrderCreated.new, stream_name)
       end
-    end
-
-    def create_event(event, stream_name)
-      facade.publish_event(event, stream_name)
     end
   end
 end