+++ ruby_event_store/spec/actions/read_all_events_spec.rb
@@ -2,16 +2,10 @@ require_relative '../spec_helper'
 
 module RubyEventStore
   describe Facade do
-
-    let(:repository)  { InMemoryRepository.new }
-    let(:facade)      { RubyEventStore::Facade.new(repository) }
     let(:stream_name) { 'stream_name' }
 
-    before(:each) do
-      repository.reset!
-    end
-
     specify 'raise exception if stream name is incorrect' do
+      facade = RubyEventStore::Facade.new(InMemoryRepository.new)
       expect { facade.read_stream_events_forward(nil) }.to raise_error(IncorrectStreamData)
       expect { facade.read_stream_events_forward('') }.to raise_error(IncorrectStreamData)
       expect { facade.read_stream_events_backward(nil) }.to raise_error(IncorrectStreamData)
@@ -19,7 +13,8 @@ module RubyEventStore
     end
 
     specify 'return all events ordered forward' do
-      prepare_events_in_store
+      facade = RubyEventStore::Facade.new(InMemoryRepository.new)
+      prepare_events_in_store(facade)
       events = facade.read_stream_events_forward(stream_name)
       expect(events[0]).to be_event({event_id: '0', event_type: 'OrderCreated', data: {}})
       expect(events[1]).to be_event({event_id: '1', event_type: 'OrderCreated', data: {}})
@@ -28,7 +23,8 @@ module RubyEventStore
     end
 
     specify 'return all events ordered backward' do
-      prepare_events_in_store
+      facade = RubyEventStore::Facade.new(InMemoryRepository.new)
+      prepare_events_in_store(facade)
       events = facade.read_stream_events_backward(stream_name)
       expect(events[0]).to be_event({event_id: '3', event_type: 'OrderCreated', data: {}})
       expect(events[1]).to be_event({event_id: '2', event_type: 'OrderCreated', data: {}})
@@ -38,15 +34,11 @@ module RubyEventStore
 
     private
 
-    def prepare_events_in_store
+    def prepare_events_in_store(facade)
       4.times do |index|
         event = OrderCreated.new(event_id: index)
-        create_event(event, stream_name)
+        facade.publish_event(event, stream_name)
       end
-    end
-
-    def create_event(event, stream_name)
-      facade.publish_event(event, stream_name)
     end
   end
 end