+++ ruby_event_store/spec/in_memory_repository_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+require 'ruby_event_store/spec/event_repository_lint'
+
+module RubyEventStore
+  describe InMemoryRepository do
+    it_behaves_like :event_repository, InMemoryRepository
+  end
+end